Moved time-series stuff into separate module
